+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - [251](https://github.com/vegaprotocol/data-node/pull/251) - Updated proto and core and added support for the new events (state var and network limits)
 
 ### 🐛 Fixes
+- [](https://github.com/vegaprotocol/data-node/pull/) -
+- [277](https://github.com/vegaprotocol/data-node/pull/277) - Now returns not-found error instead of internal error when proposal not found 
 - [274](https://github.com/vegaprotocol/data-node/issues/274) - Bug fix for proposal NO vote showing incorrect weight and tokens